+++ CSharpExcelChangeLogger/ChangeLogger/Memory/IMemoryComparison.cs
@@ -46,6 +46,26 @@ namespace CSharpExcelChangeLogger.ChangeLogger.Memory
         bool LocationMatchesAndDataMatches { get; }
 
         /// <summary>
+        /// Address of the range stored in memory before the change occurred.
+        /// </summary>
+        string? RangeAddressBeforeChange { get; }
+
+        /// <summary>
+        /// Address of the range where the change occurred.
+        /// </summary>
+        string? RangeAddressAfterChange { get; }
+
+        /// <summary>
+        /// Name of the sheet data stored in memory before the change occurred.
+        /// </summary>
+        string? SheetNameBeforeChange { get; }
+
+        /// <summary>
+        /// Name of the sheet where the change occurred.
+        /// </summary>
+        string? SheetNameAfterChange { get; }
+
+        /// <summary>
         /// Data saved to memory before the change occurred. Will be null if no data was stored in memory.
         /// </summary>
         string[,]? DataBeforeChange { get; }