+++ CSharpExcelChangeLogger/ChangeLogger/Memory/MemoryComparison.cs
@@ -1,4 +1,6 @@
-﻿namespace CSharpExcelChangeLogger.ChangeLogger.Memory
+﻿using System;
+
+namespace CSharpExcelChangeLogger.ChangeLogger.Memory
 {
     class MemoryComparison : IMemoryComparison
     {
@@ -16,14 +18,30 @@
 
         public bool LocationMatchesAndDataMatches => LocationMatches && DataMatches;
 
+        public string? RangeAddressBeforeChange { get; }
+
+        public string? RangeAddressAfterChange { get; }
+
+        public string? SheetNameBeforeChange { get; }
+
+        public string? SheetNameAfterChange { get; }
+
         public string[,]? DataBeforeChange { get; }
 
         public string[,]? DataAfterChange { get; }
 
-        public MemoryComparison(
-            bool locationMatches, bool dataMatches, 
-            bool isNewRow, bool isRowDelete, bool isNewColumn, bool isColumnDelete, 
-            string[,]? dataBeforeChange, string[,]? dataAfterChange)
+        public MemoryComparison(bool locationMatches,
+                                bool dataMatches,
+                                bool isNewRow,
+                                bool isRowDelete,
+                                bool isNewColumn,
+                                bool isColumnDelete,
+                                string? rangeAddressBeforeChange,
+                                string? rangeAddressAfterChange,
+                                string? sheetNameBeforeChange,
+                                string? sheetNameAfterChange,
+                                string[,]? dataBeforeChange,
+                                string[,]? dataAfterChange)
         {
             LocationMatches = locationMatches;
             DataMatches = dataMatches;
@@ -31,6 +49,10 @@
             IsRowDelete = isRowDelete;
             IsNewColumn = isNewColumn;
             IsColumnDelete = isColumnDelete;
+            RangeAddressBeforeChange = rangeAddressBeforeChange;
+            RangeAddressAfterChange = rangeAddressAfterChange;
+            SheetNameBeforeChange = sheetNameBeforeChange;
+            SheetNameAfterChange = sheetNameAfterChange;
             DataBeforeChange = dataBeforeChange;
             DataAfterChange = dataAfterChange;
         }