+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Rename DoesMemoryMatch method to Compare (#22)
 
+### Added
+- Add properties on the memory comparison object for the range address and sheet name before and after a change (#24)
 
 ## [0.1.0] - 2020-05-18